outer_space improvement

+++ outer_space.sql
@@ -11,7 +11,7 @@ CREATE TABLE planets
 (
   id SERIAL PRIMARY KEY,
   name TEXT NOT NULL,
-  orbital_period_in_years FLOAT NOT NULL,
+  orbital_period_in_years INTEGER NOT NULL,
   orbits_around TEXT NOT NULL,
   galaxy TEXT NOT NULL,
   moons TEXT[]
